- | Delete the version of CAN_validator that will not be used.
